final fix

+++ test/FullLifecycle.t.sol
@@ -11,7 +11,7 @@ contract FullLifecycleTest is Test {
     PolicastMarketV3 public market;
     PolicastViews public views;
     MockERC20 public token;
-    
+
     // 10 users for comprehensive testing
     address public creator = address(0x1234567890123456789012345678901234567890);
     address public user1 = address(0x1111111111111111111111111111111111111111);
@@ -24,92 +24,92 @@ contract FullLifecycleTest is Test {
     address public user8 = address(0x8888888888888888888888888888888888888888);
     address public user9 = address(0x9999999999999999999999999999999999999999);
     address public user10 = address(0xaAaAaAaaAaAaAaaAaAAAAAAAAaaaAaAaAaaAaaAa);
-    
+
     address[] public allUsers;
     uint256 public marketId;
-    
+
     function setUp() public {
         // Deploy contracts
         token = new MockERC20(10000000e18); // 10M total supply
-        
+
         vm.prank(creator);
         market = new PolicastMarketV3(address(token));
         views = new PolicastViews(address(market));
-        
+
         // Set up all users array
         allUsers = [user1, user2, user3, user4, user5, user6, user7, user8, user9, user10];
-        
+
         // Grant necessary roles to creator
         vm.startPrank(creator);
         market.grantQuestionCreatorRole(creator);
         market.grantMarketValidatorRole(creator);
         market.grantQuestionResolveRole(creator);
         vm.stopPrank();
-        
+
         // Give creator initial tokens for market creation
         token.transfer(creator, 2000000e18); // 2M tokens (enough for 200k liquidity)
         vm.prank(creator);
         token.approve(address(market), type(uint256).max);
-        
+
         // Give each user substantial tokens for trading
-        for (uint i = 0; i < allUsers.length; i++) {
+        for (uint256 i = 0; i < allUsers.length; i++) {
             token.transfer(allUsers[i], 100000e18); // 100k tokens each
             vm.prank(allUsers[i]);
             token.approve(address(market), type(uint256).max);
         }
     }
-    
+
     function testFullMarketLifecycle() public {
         console.log("\n=== POLICAST FULL MARKET LIFECYCLE TEST ===");
         console.log("Testing 1:100 share-to-token ratio with 10 users");
         console.log("Each winning share pays exactly 100 tokens\n");
-        
+
         // 1. Create Market
         _createMarket();
-        
+
         // 2. Show initial prices
         _showInitialPrices();
-        
+
         // 3. Phase 1: Initial buying by various users
         console.log("=== PHASE 1: INITIAL BUYING ===");
         _phase1_InitialBuying();
-        
+
         // 4. Phase 2: Some users sell to take profits
         console.log("\n=== PHASE 2: SELLING AND PRICE MOVEMENTS ===");
         _phase2_SellingPhase();
-        
-        // 5. Phase 3: Final trading before resolution  
+
+        // 5. Phase 3: Final trading before resolution
         console.log("\n=== PHASE 3: FINAL TRADING RUSH ===");
         _phase3_FinalTrading();
-        
+
         // 6. Show final pre-resolution state
         console.log("\n=== PRE-RESOLUTION SUMMARY ===");
         _showPreResolutionState();
-        
+
         // 7. Resolve market
         console.log("\n=== MARKET RESOLUTION ===");
         _resolveMarket();
-        
+
         // 8. Users claim winnings
         console.log("\n=== CLAIMING WINNINGS ===");
         _claimWinnings();
-        
+
         // 9. Final summary
         console.log("\n=== FINAL SUMMARY ===");
         _finalSummary();
     }
-    
+
     function _createMarket() internal {
         string[] memory options = new string[](3);
         options[0] = "Bitcoin (BTC)";
         options[1] = "Ethereum (ETH)";
         options[2] = "Solana (SOL)";
-        
+
         string[] memory symbols = new string[](3);
         symbols[0] = "BTC";
         symbols[1] = "ETH";
         symbols[2] = "SOL";
-        
+
         vm.prank(creator);
         marketId = market.createMarket(
             "Which crypto will perform best this quarter?",
@@ -122,16 +122,16 @@ contract FullLifecycleTest is Test {
             200000e18, // 200k tokens initial liquidity (enough to cover worst-case payouts)
             false // not early resolution
         );
-        
+
         vm.prank(creator);
         market.validateMarket(marketId);
-        
+
         // CRITICAL: Ensure contract has enough tokens to cover worst-case payouts
         // In a real scenario, the initial liquidity should be sized appropriately
         // For this test, we'll add extra tokens to guarantee all winners can claim
         uint256 maxPossiblePayout = 30 * 100e18; // 30 users × 100 tokens/share (conservative)
         token.transfer(address(market), maxPossiblePayout);
-        
+
         console.log("Market Created:");
         console.log("- Market ID:");
         console.log(marketId);
@@ -140,11 +140,11 @@ contract FullLifecycleTest is Test {
         console.log("- Initial liquidity: 200,000 tokens");
         console.log("- Additional payout reserve: 3,000 tokens");
     }
-    
+
     function _showInitialPrices() internal view {
         console.log("\nInitial Prices (probabilities):");
         string[3] memory options = ["BTC", "ETH", "SOL"];
-        for (uint i = 0; i < 3; i++) {
+        for (uint256 i = 0; i < 3; i++) {
             uint256 price = views.calculateCurrentPrice(marketId, i);
             console.log("-");
             console.log(options[i]);
@@ -152,56 +152,56 @@ contract FullLifecycleTest is Test {
             console.log((price * 100) / 1e18);
         }
     }
-    
+
     function _phase1_InitialBuying() internal {
         // User1: Bullish on BTC, buys 5 shares
         _executeTrade(user1, "User1", 0, "BTC", 5e18, true);
-        
-        // User2: Believes in ETH, buys 3 shares  
+
+        // User2: Believes in ETH, buys 3 shares
         _executeTrade(user2, "User2", 1, "ETH", 3e18, true);
-        
+
         // User3: SOL maximalist, buys 4 shares
         _executeTrade(user3, "User3", 2, "SOL", 4e18, true);
-        
+
         // User4: Hedges with BTC, buys 2 shares
         _executeTrade(user4, "User4", 0, "BTC", 2e18, true);
-        
+
         // User5: Big ETH bet, buys 6 shares
         _executeTrade(user5, "User5", 1, "ETH", 6e18, true);
-        
+
         _showCurrentPrices("After Phase 1");
     }
-    
+
     function _phase2_SellingPhase() internal {
         // User1: Takes some profit, sells 2 BTC shares
         _executeTrade(user1, "User1", 0, "BTC", 2e18, false);
-        
+
         // User2: Doubles down on ETH instead of selling
         _executeTrade(user2, "User2", 1, "ETH", 2e18, true);
-        
+
         // User6: New player, buys SOL
         _executeTrade(user6, "User6", 2, "SOL", 3e18, true);
-        
+
         // User3: Sells half SOL position
         _executeTrade(user3, "User3", 2, "SOL", 2e18, false);
-        
+
         _showCurrentPrices("After Phase 2");
     }
-    
+
     function _phase3_FinalTrading() internal {
         // Final rush of trading
         _executeTrade(user7, "User7", 0, "BTC", 1e18, true);
         _executeTrade(user8, "User8", 1, "ETH", 2e18, true);
         _executeTrade(user9, "User9", 2, "SOL", 1e18, true);
         _executeTrade(user10, "User10", 1, "ETH", 4e18, true);
-        
+
         // Some users change positions
         _executeTrade(user5, "User5", 1, "ETH", 1e18, false);
         _executeTrade(user4, "User4", 0, "BTC", 1e18, true);
-        
+
         _showCurrentPrices("After Phase 3");
     }
-    
+
     function _executeTrade(
         address user,
         string memory userName,
@@ -238,12 +238,12 @@ contract FullLifecycleTest is Test {
             console.log("tokens");
         }
     }
-    
+
     function _showCurrentPrices(string memory phase) internal view {
         console.log(phase);
         console.log("- Current Prices:");
         string[3] memory options = ["BTC", "ETH", "SOL"];
-        for (uint i = 0; i < 3; i++) {
+        for (uint256 i = 0; i < 3; i++) {
             uint256 price = views.calculateCurrentPrice(marketId, i);
             console.log("-");
             console.log(options[i]);
@@ -251,17 +251,17 @@ contract FullLifecycleTest is Test {
             console.log((price * 100) / 1e18);
         }
     }
-    
+
     function _showPreResolutionState() internal view {
         string[3] memory options = ["BTC", "ETH", "SOL"];
         console.log("User Holdings Before Resolution:");
-        
-        for (uint userIdx = 0; userIdx < allUsers.length; userIdx++) {
+
+        for (uint256 userIdx = 0; userIdx < allUsers.length; userIdx++) {
             address user = allUsers[userIdx];
             string memory userName = _getUserName(userIdx);
-            
+
             bool hasShares = false;
-            for (uint optionIdx = 0; optionIdx < 3; optionIdx++) {
+            for (uint256 optionIdx = 0; optionIdx < 3; optionIdx++) {
                 uint256 shares = market.getMarketOptionUserShares(marketId, optionIdx, user);
                 if (shares > 0) {
                     if (!hasShares) {
@@ -276,45 +276,45 @@ contract FullLifecycleTest is Test {
                 }
             }
         }
-        
+
         _showCurrentPrices("Final");
     }
-    
+
     function _resolveMarket() internal {
         // Fast forward past the market end time
         vm.warp(block.timestamp + 8 days);
-        
+
         // For testing assume ETH (optionId 1) is the winner
         vm.prank(creator);
         market.resolveMarket(marketId, 1);
         console.log("Market resolved to ETH (optionId 1)");
     }
-    
+
     function _claimWinnings() internal {
         console.log("Winning users claiming their rewards:");
-        
-        for (uint userIdx = 0; userIdx < allUsers.length; userIdx++) {
+
+        for (uint256 userIdx = 0; userIdx < allUsers.length; userIdx++) {
             address user = allUsers[userIdx];
             string memory userName = _getUserName(userIdx);
-            
+
             // Check if user has winning shares (ETH shares)
             uint256 winningShares = market.getMarketOptionUserShares(marketId, 1, user);
             if (winningShares > 0) {
                 uint256 balanceBefore = token.balanceOf(user);
-                
+
                 vm.prank(user);
                 market.claimWinnings(marketId);
-                
+
                 uint256 balanceAfter = token.balanceOf(user);
                 uint256 payout = balanceAfter - balanceBefore;
-                
+
                 console.log(userName);
                 console.log("claimed");
                 console.log(payout / 1e18);
                 console.log("tokens for");
                 console.log(winningShares / 1e18);
                 console.log("ETH shares");
-                
+
                 // additional sanity log: tokens per share (integer division)
                 if (winningShares > 0) {
                     console.log("tokens per share (approx):");
@@ -323,7 +323,7 @@ contract FullLifecycleTest is Test {
             }
         }
     }
-    
+
     function _finalSummary() internal view {
         console.log("=== 1:100 RATIO DEMONSTRATION COMPLETE ===");
         console.log("Key observations:");
@@ -332,21 +332,21 @@ contract FullLifecycleTest is Test {
         console.log("3. Each winning share paid exactly 100 tokens");
         console.log("4. LMSR price movements worked correctly with 100x scaling");
         console.log("5. Market feels substantial and engaging for traders");
-        
+
         // Verify the 100 tokens per share payout
         uint256 totalEthShares = 0;
-        for (uint userIdx = 0; userIdx < allUsers.length; userIdx++) {
+        for (uint256 userIdx = 0; userIdx < allUsers.length; userIdx++) {
             uint256 shares = market.getMarketOptionUserShares(marketId, 1, allUsers[userIdx]);
             totalEthShares += shares;
         }
-        
+
         console.log("Total ETH shares held:");
         console.log(totalEthShares / 1e18);
         console.log("Expected total payout:");
         console.log((totalEthShares / 1e18) * 100);
         console.log("tokens");
     }
-    
+
     function _getUserName(uint256 userIdx) internal pure returns (string memory) {
         if (userIdx == 0) return "User1";
         if (userIdx == 1) return "User2";
